backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.10.4
python-multipart==0.0.20
cryptography==45.0.4
